refactor: extract input and execution of longname to its own feature
